Refactor : move type.ts
