+++ src/app/(logged)/layout.tsx
@@ -3,7 +3,7 @@ import type { Metadata } from 'next'
 import '../../styles/main.css'
 import '../globals.css'
 import { redirect } from 'next/navigation'
-import { UserProfile } from '@/types/user'
+import { UserProfile } from '@/types/types'
 import Header from '@/containers/header'
 import RecoilRootWrapper from '@/containers/recoilRootWrapper'
 import { KAKAO_MAP_KEY } from '@/constants/contant'
