--- types.ts
@@ -1,6 +1,0 @@
-
-export interface AddressType {
-  address: string;
-  postCode: string;
-  additional: string;
-}
